feat: Update operator-sdk and align Makefiles (#21)
Updates to operator-sdk v1.39.2 and aligns the Makefiles for better
consistency with operator-sdk generated Makefile. It supersedes #20, and
allows for a more elegant way for installing tools locally (like its
done nowadays in the operator-sdk generated Makefiles).

Also aligned versions with the versions that operator-sdk v1.39.2 is
using.

Changes include:
- Update operator-sdk to v1.39.2
- Align Makefile with latest operator-sdk scaffold
- Improve tool dependency management
- Add pre-commit makefile target
- Update dependencies in go.mod
- Refactor install procedures for tools

Note: Due to the way how tools are downloaded now, it's recommended to
remove `bin/` to remove legacy tools.