feat(kustomizer): add kustomize plugin support for manifest rendering (#89)
This change refactors the Kustomize component to support a flexible plugin-based architecture, which separates the concerns of rendering manifests from applying them to the cluster.

This new design is more scalable and maintainable, as new functionality can now be added via self-contained plugins without modifying the core rendering engine. It also prevents a regression in existing functionality (like name prefixing and namespace injection), which have been reimplemented as plugins.


Approved-by: rhdedgar

Approved-by: VaishnaviHire