Revert "chore: Add LICENSE" (#103)
Reverts llamastack/llama-stack-k8s-operator#98

Sorry, I was a bit premature with my approval, as we probably still need a bit more discussion about licensing in general and especially about ownership. IANAL, but let's keep the PR open for a bit longer to find a good solution.

So I will revert the previous PR if okay, but at the same time, I will open an identical one to llamastack/llama-stack-k8s-operator#98 to continue the discussion.


Approved-by: leseb