Adding MIT License (#104)
This is a re-make of #98 , that was prematurely approved (sorry, my bad).

People have still some concerns that we need to resolve before merging the license. 
It can be an easy discussion, but IANAL, so better play safe and listen to some more voices.

The questions actually are:

* What is the implication to have `Copyright (c) 2025 Meta Llama` wrt to future work and when moving potentially to a foundation later ? Wouldn't it be bettet to use a neutral term like `Copyright (c) 2025 The Llama Stack community` ?
* There are still repos in the GitHub `llamastack` org that uses an APL license, eg. [llama-stack-client-typescript/LICENSE](https://github.com/llamastack/llama-stack-client-typescript/blob/generated/LICENSE), but I have the feeling all repos in that org should leverage the same license. 
* Also, other projects that are using the MIT license, have a different copyright holder line, like `Copyright (c) Meta Platforms, Inc. and affiliates` in [llamastack/llama-stack-client-python](https://github.com/llamastack/llama-stack-client-python/blob/fe7130c105885a97cbc67410050a45f3de2f2db5/LICENSE)

My proposal would be to settle on a single license file with the same content and update all repositories atomically.




Approved-by: leseb

Approved-by: VaishnaviHire

Approved-by: nathan-weinberg